Make app compile again

+++ Code/macOS/Configuration/Style/Float+Style.swift
@@ -1,23 +1,23 @@
 import SwiftyToolz
 
-extension Float
+extension Double
 {
-    static let groupIconWidthFactor: Float = 0.75
-    static let relativeTextInset: Float = 0.85
-    static let listCornerRadius: Float = 6
-    static let listGap: Float = 10
-    static let cornerRadius: Float = 3
-    static let progressBarHeight: Float = 10
+    static let groupIconWidthFactor: Double = 0.75
+    static let relativeTextInset: Double = 0.85
+    static let listCornerRadius: Double = 6
+    static let listGap: Double = 10
+    static let cornerRadius: Double = 3
+    static let progressBarHeight: Double = 10
     
-    static func itemPadding(for lineHeight: Float) -> Float
+    static func itemPadding(for lineHeight: Double) -> Double
     {
         0.5 * lineHeight + lineSpacing(for: lineHeight)
     }
     
-    static func lineSpacing(for lineHeight: Float) -> Float
+    static func lineSpacing(for lineHeight: Double) -> Double
     {
-        Float(Int(relativeSpacing * lineHeight + 0.5))
+        Double(Int(relativeSpacing * lineHeight + 0.5))
     }
     
-    private static let relativeSpacing: Float = 5.0 / 17.0 // 5 @ line height 17
+    private static let relativeSpacing: Double = 5.0 / 17.0 // 5 @ line height 17
 }
